moving from loading screen to inline loading indicator
