+++ packages/mobile/src/paymentRequest/IncomingPaymentRequestListItem.tsx
@@ -1,10 +1,11 @@
 import ContactCircle from '@celo/react-components/components/ContactCircle'
 import RequestMessagingCard from '@celo/react-components/components/RequestMessagingCard'
 import BigNumber from 'bignumber.js'
-import * as React from 'react'
+import React, { useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { StyleSheet, View } from 'react-native'
-import { useDispatch } from 'react-redux'
+import { ActivityIndicator, StyleSheet, View } from 'react-native'
+import { useDispatch, useSelector } from 'react-redux'
+import { errorSelector } from 'src/alert/reducer'
 import { HomeEvents } from 'src/analytics/Events'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import { TokenTransactionType } from 'src/apollo/types'
@@ -13,6 +14,13 @@ import { declinePaymentRequest } from 'src/firebase/actions'
 import { CURRENCIES, CURRENCY_ENUM } from 'src/geth/consts'
 import { NotificationBannerCTATypes, NotificationBannerTypes } from 'src/home/NotificationBox'
 import { Namespaces } from 'src/i18n'
+import { fetchAddressesAndValidate } from 'src/identity/actions'
+import {
+  AddressValidationType,
+  isFetchingAddressesSelector,
+  secureSendPhoneNumberMappingSelector,
+} from 'src/identity/reducer'
+import { getAddressValidationType } from 'src/identity/secureSend'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { getRecipientThumbnail, Recipient } from 'src/recipients/recipient'
@@ -29,6 +37,11 @@ interface Props {
 export default function IncomingPaymentRequestListItem({ id, amount, comment, requester }: Props) {
   const { t } = useTranslation(Namespaces.paymentRequestFlow)
   const dispatch = useDispatch()
+  const secureSendPhoneNumberMapping = useSelector(secureSendPhoneNumberMappingSelector)
+  const isFetchingAddresses = useSelector(isFetchingAddressesSelector)
+  const error = useSelector(errorSelector)
+  const prevIsFetchingAddressesRef = useRef(isFetchingAddresses)
+  const [isLoading, setIsLoading] = useState(false)
 
   const transactionData: TransactionDataInput = {
     reason: comment,
@@ -37,14 +50,23 @@ export default function IncomingPaymentRequestListItem({ id, amount, comment, re
     type: TokenTransactionType.PayRequest,
     firebasePendingRequestUid: id,
   }
+  const { recipient } = transactionData
+  const { e164PhoneNumber } = recipient
+  const addressValidationType = getAddressValidationType(recipient, secureSendPhoneNumberMapping)
 
   const onPay = () => {
+    if (e164PhoneNumber) {
+      setIsLoading(true)
+      // Need to check latest mapping to prevent user from accepting fradulent requests
+      dispatch(fetchAddressesAndValidate(e164PhoneNumber))
+    } else {
+      navigateToNextScreen()
+    }
+
     ValoraAnalytics.track(HomeEvents.notification_select, {
       notificationType: NotificationBannerTypes.incoming_tx_request,
       selectedAction: NotificationBannerCTATypes.pay,
     })
-
-    navigate(Screens.AddressFetchLoading, { transactionData })
   }
 
   const onPaymentDecline = () => {
@@ -56,19 +78,43 @@ export default function IncomingPaymentRequestListItem({ id, amount, comment, re
     Logger.showMessage(t('requestDeclined'))
   }
 
-  const name = requester.displayName
-  const requestAmount = {
-    value: amount,
-    currencyCode: CURRENCIES[CURRENCY_ENUM.DOLLAR].code,
+  const navigateToNextScreen = () => {
+    if (
+      addressValidationType === AddressValidationType.FULL ||
+      addressValidationType === AddressValidationType.PARTIAL
+    ) {
+      navigate(Screens.ValidateRecipientIntro, { transactionData, addressValidationType })
+    } else {
+      navigate(Screens.SendConfirmation, { transactionData })
+    }
   }
+
+  React.useEffect(() => {
+    const prevIsFetchingAddresses = prevIsFetchingAddressesRef.current
+    prevIsFetchingAddressesRef.current = isFetchingAddresses
+
+    if (prevIsFetchingAddresses === true && isFetchingAddresses === false) {
+      setIsLoading(false)
+      if (!error) {
+        navigateToNextScreen()
+      }
+    }
+  }, [isFetchingAddresses])
 
   return (
     <View style={styles.container}>
       <RequestMessagingCard
         testID={`IncomingPaymentRequestNotification/${id}`}
-        title={t('incomingPaymentRequestNotificationTitle', { name })}
+        title={t('incomingPaymentRequestNotificationTitle', { name: requester.displayName })}
         details={comment}
-        amount={<CurrencyDisplay amount={requestAmount} />}
+        amount={
+          <CurrencyDisplay
+            amount={{
+              value: amount,
+              currencyCode: CURRENCIES[CURRENCY_ENUM.DOLLAR].code,
+            }}
+          />
+        }
         icon={
           <ContactCircle
             address={requester.address}
@@ -78,7 +124,11 @@ export default function IncomingPaymentRequestListItem({ id, amount, comment, re
         }
         callToActions={[
           {
-            text: t('global:send'),
+            text: isLoading ? (
+              <ActivityIndicator testID={'loading/paymentRequest'} />
+            ) : (
+              t('global:send')
+            ),
             onPress: onPay,
           },
           {
